+++ srcs/ft_strnchr.c
@@ -1,28 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   libft.h                                            :+:      :+:    :+:   */
+/*   ft_strnchr.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: nihuynh <nihuynh@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2018/03/30 10:34:50 by nihuynh           #+#    #+#             */
-/*   Updated: 2018/04/03 16:33:50 by nihuynh          ###   ########.fr       */
+/*   Created: 2018/04/29 15:19:42 by nihuynh           #+#    #+#             */
+/*   Updated: 2018/04/29 15:20:07 by nihuynh          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef LIBFT_H
-# define LIBFT_H
+#include "ftstring.h"
 
-# include "ftconvert.h"
-# include "ftctype.h"
-# include "ftio.h"
-# include "ftlist.h"
-# include "ftmath.h"
-# include "ftmem.h"
-# include "ftstring.h"
-# include "ftgnl.h"
-
-int		ft_tablen(char **tab);
-void	ft_tabdel(char ***tab);
-
-#endif
+char	*ft_strnchr(char const *str, int c, int size)
+{
+	while ((*str == (char)c || *str) && size--)
+	{
+		if (*str == (char)c)
+			return ((char *)str);
+		str++;
+	}
+	return (NULL);
+}